section_13_B_AuthJS_AutenProct.Rutas_PrismaAdapter_BBDD_Final

+++ src/app/api/todos/route.ts
@@ -3,64 +3,77 @@ import { NextResponse, NextRequest } from 'next/server'
 import { argv } from 'process';
 import * as yup from 'yup';
 import { PrismaClientOptions } from '../../../generated/prisma/internal/prismaNamespace';
+import { getUserSessionServer } from '@/auth/components/actions/auth-actions';
 
-export async function GET(request: Request) { 
-  
-const {searchParams } =new URL(request.url);
-const take = Number( searchParams.get('take') ?? '10');
-const skip = Number( searchParams.get('skip') ?? '0');
+export async function GET(request: Request) {
 
-  if ( isNaN(take) ){ 
+  const { searchParams } = new URL(request.url);
+  const take = Number(searchParams.get('take') ?? '10');
+  const skip = Number(searchParams.get('skip') ?? '0');
+
+  if (isNaN(take)) {
     return NextResponse.json({ message: 'Take tiene que ser un número' }, { status: 400 });
   }
 
-  if ( isNaN(skip) ){ 
+  if (isNaN(skip)) {
     return NextResponse.json({ message: 'Skip tiene que ser un número' }, { status: 400 });
   }
 
 
   const todos = await prisma.todo.findMany({
-    take : take,  //+take ,+(take)
+    take: take,  //+take ,+(take)
     skip: skip,
   });
 
-  
-  return NextResponse.json( todos );
+
+  return NextResponse.json(todos);
 }
 
 
 const postSchema = yup.object({
-    description: yup.string().required(),
-    complete: yup.boolean().optional().default(false),
+  description: yup.string().required(),
+  complete: yup.boolean().optional().default(false),
 })
 
 
-export async function POST (request: Request){
+export async function POST(request: Request) {
 
-    try {
+  const user = await getUserSessionServer()
 
-        const {complete, description} = await postSchema.validate( await request.json());
+  if (!user) {
+    return NextResponse.json('No autorizado', { status: 401 })
+  }
 
-        const todo = await prisma.todo.create({ data: {complete,description}})
+  try {
 
-        return NextResponse.json(todo);
-        
-    } catch (error) {
-        return NextResponse.json(error,{status: 400});
-    }
-    
+    const { complete, description } = await postSchema.validate(await request.json());
+
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
+
+    return NextResponse.json(todo);
+
+  } catch (error) {
+    return NextResponse.json(error, { status: 400 });
+  }
+
 }
 
-export async function DELETE (request: Request){
+export async function DELETE(request: Request) {
 
-    try {
+  const user = await getUserSessionServer()
 
-        await prisma.todo.deleteMany({ where:{complete: true}})
+  if (!user) {
+    return NextResponse.json('No autorizado', { status: 401 })
+  }
 
-        return NextResponse.json('borrados');
-        
-    } catch (error) {
-        return NextResponse.json(error,{status: 400});
-    }
-    
+  try {
+
+    await prisma.todo.deleteMany({ where: { complete: true,userId:user.id } })
+
+    return NextResponse.json('borrados');
+
+  } catch (error) {
+    return NextResponse.json(error, { status: 400 });
+  }
+
 }
